requirements_simple.txt
# Simplified requirements for BookSpine Recommender (without Goodreads API)
# Core FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Environment
python-dotenv==1.0.0

# HTTP requests
requests==2.31.0

# Image processing
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3

# Machine Learning
ultralytics==8.0.196

# OpenAI API
openai==1.3.5

# File handling
aiofiles==23.2.1

# Validation
pydantic==2.5.0
